fix: simplify base processor code
